Migrate hw-reset/test-stress.py to pytest
